+++ tests/unit_tests/test_chat_models.py
@@ -5,7 +5,7 @@ from typing import Type
 from dotenv import load_dotenv
 from langchain_tests.unit_tests import ChatModelUnitTests
 
-from langchain_qwq.chat_models import ChatQwQ
+from langchain_qwq.chat_models import ChatQwen, ChatQwQ
 
 load_dotenv()
 
@@ -20,5 +20,19 @@ class TestChatQwQUnit(ChatModelUnitTests):
         # These should be parameters used to initialize your integration for testing
         return {
             "model": "qwq-plus",
+            "temperature": 0,
+        }
+
+
+class TestChatQwenUnit(ChatModelUnitTests):
+    @property
+    def chat_model_class(self) -> Type[ChatQwen]:
+        return ChatQwen
+
+    @property
+    def chat_model_params(self) -> dict:
+        # These should be parameters used to initialize your integration for testing
+        return {
+            "model": "qwen-plus-latest",
             "temperature": 0,
         }
